Dump Stargate logs on API check failure for all tests
